Lesson2_Task1

+++ Lesson2/PriceAndDiscount/Program.cs
@@ -11,7 +11,7 @@ namespace PriceAndDiscount
         static void Main(string[] args)
         {
             Console.WriteLine("Enter the price, please: ");
-            double price = Convert.ToInt32(Console.ReadLine());
+            double price = Convert.ToDouble(Console.ReadLine());
             Console.WriteLine("Please enter the discount percentage: ");
             int discountInPercent = Convert.ToInt32(Console.ReadLine());
